Fix event conversion to common schema
